AdaptiveTestPlots.jl

This is a package for plotting recordings of CATs (Computerised Adaptive Tests).

Index

AdaptiveTestPlots.AdaptiveTestPlots

AdaptiveTestPlots.CatRecorder

AdaptiveTestPlots.CatRecorder

AdaptiveTestPlots.LabelledToggleGrid

AdaptiveTestPlots.MenuGrid

AdaptiveTestPlots.ability_convergence_lines

AdaptiveTestPlots.ability_evolution_lines

AdaptiveTestPlots.lh_evolution_interactive

AdaptiveTestPlots.plot_item_bank

AdaptiveTestPlots.plot_item_bank_comparison

AdaptiveTestPlots.plot_likelihoods

API

AdaptiveTestPlots.AdaptiveTestPlots — Module

This module contains helpers for creating CAT/IRT related plots. This module requires the optional depedencies AlgebraOfGraphics, DataFrames and Makie to be installed.

AdaptiveTestPlots.CatRecorder — Type

CatRecorder(
    xs,
    points,
    ability_ests,
    num_questions,
    num_respondents,
    integrator,
    raw_estimator,
    ability_estimator
) -> CatRecorder
CatRecorder(
    xs,
    points,
    ability_ests,
    num_questions,
    num_respondents,
    integrator,
    raw_estimator,
    ability_estimator,
    actual_abilities
) -> CatRecorder

AdaptiveTestPlots.CatRecorder — Method

AdaptiveTestPlots.LabelledToggleGrid — Type

AdaptiveTestPlots.LabelledToggleGrid <: Block

No docstring defined.

Attributes

(type ?AdaptiveTestPlots.LabelledToggleGrid.x in the REPL for more information about attribute x)

alignmode, halign, height, tellheight, tellwidth, valign, width

AdaptiveTestPlots.MenuGrid — Type

AdaptiveTestPlots.MenuGrid <: Block

No docstring defined.

Attributes

(type ?AdaptiveTestPlots.MenuGrid.x in the REPL for more information about attribute x)

alignmode, halign, height, tellheight, tellwidth, valign, width

AdaptiveTestPlots.ability_convergence_lines — Method

ability_convergence_lines(recorder; abilities)

AdaptiveTestPlots.ability_evolution_lines — Method

ability_evolution_lines(
    recorder;
    abilities
) -> AlgebraOfGraphics.FigureGrid

AdaptiveTestPlots.lh_evolution_interactive — Method

lh_evolution_interactive(
    recorder;
    abilities
) -> Makie.Figure

AdaptiveTestPlots.plot_item_bank — Method

plot_item_bank(
    item_bank::FittedItemBanks.AbstractItemBank;
    items,
    labeller,
    zero_symmetric,
    include_outcome_toggles,
    include_item_toggles
) -> Makie.Figure

Plot an item bank item_bank with items items using the labeller labeller to label the items.

Lines are drawn for each item and each outcome. The domain of the item bank is used to determine the x-axis limits. You can use zero_symmetric to force the domain to be symmetric about zero.

If include_outcome_toggles is true, then a toggle grid is drawn to show/hide the outcomes. If include_item_toggles is true, then a toggle grid is drawn to show/hide the items.

AdaptiveTestPlots.plot_item_bank_comparison — Method

plot_item_bank_comparison(
    item_banks::AbstractVector;
    items,
    labeller,
    include_outcome_toggles,
    include_item_toggles,
    ignore_domain_indices
) -> Makie.Figure

Plot a comparison of multiple item banks item_banks. For an explanation of the options, see: plot_item_bank.

AdaptiveTestPlots.plot_likelihoods — Function

plot_likelihoods(
    estimators,
    tracked_responses,
    integrator,
    xs
) -> Makie.Figure
plot_likelihoods(
    estimators,
    tracked_responses,
    integrator,
    xs,
    lim_lo
) -> Makie.Figure
plot_likelihoods(
    estimators,
    tracked_responses,
    integrator,
    xs,
    lim_lo,
    lim_hi
) -> Makie.Figure